backend/requirements.txt
langchain
flask
requests
pymongo
python-dotenv
langchain-openai
flask_cors